requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
billiard==4.2.0
boto3==1.34.66
botocore==1.34.66
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
contextlib2==21.6.0
cryptography==42.0.2
dataclasses-json==0.6.6
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
Django==5.0.1
django-celery-results==2.5.1
django-cors-headers
django-cors-middleware
django-ninja
django-ninja-extra
django-ninja-jwt
django-storages
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.13.1
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanize==4.9.0
idna==3.6
img2pdf==0.5.1
injector==0.21.0
isodate==0.6.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.57
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
ocrmypdf==16.2.0
openai==1.14.2
orjson==3.10.3
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pdfminer.six==20231228
pi_heif==0.16.0
pikepdf==8.15.1
pillow==10.2.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
# psycopg2>=2.8.6,<2.9#psycopg2==2.9.9
psycopg2-binary
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
pdfplumber
# PyMuPDF==1.21.1
# PyMuPDFb==1.23.22
pyparsing==3.1.2
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
redis==5.0.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
# scikit-learn==1.4.0
scipy==1.12.0
semchunk==0.2.3
# sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.4.4
sympy==1.12
tenacity==8.3.0
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
# underthesea==6.8.0
# underthesea_core==1.0.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4

setuptools
pyvi

uWSGI>=2.0.19.1,<2.1